backend/requirements.txt
  fastapi==0.116.1
  uvicorn
  python-multipart==0.0.20
  boto3==1.35.0
  python-dotenv==1.0.0
  sqlalchemy==2.0.23
  psycopg2-binary==2.9.7
  alembic==1.13.1
  asyncpg==0.29.0